requirements.txt
absl-py==2.3.1
appdirs==1.4.4
asgiref==3.10.0
astunparse==1.6.3
atomicwrites==1.4.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
graphviz==0.21
grpcio==1.75.1
jsonpointer==2.1
keras==3.11.3
libclang==18.1.1
mkl-service==2.4.0
ml-dtypes==0.3.2
namex==0.1.0
numpy==1.26.4
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
protobuf==3.20.3
pyasn1-modules==0.2.8
pyls-spyder==0.4.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
setuptools==69.5.1
sqlparse==0.5.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==3.1.0
typing_extensions==4.15.0
wheel==0.43.0
gunicorn
psycopg2-binary
whitenoise
